refactor(backend): extract test configuration crud
Continues the incremental split of the crud monolith. TestConfiguration
moves to its own module; callers switch to direct submodule imports.
Pure move -- no logic changes.
